refactor: move GraphQLQueryEditor code out of Editor
